Lightly refactored the StigmaExecutor to separate out the task of getting files.
There are NOT duplicates in the list of gathered files (from the glob).  The problem
is that many paths contains "$1" which in bash is one of the reserved positional parameters used
for command line arguments to a bash script.  It is null on the command line.  Anyway, when a path
containing $1 (or some other similar strings) it is replaced by null  / empty string and
THE WRONG FILE IS PROCESSED!